+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Version 1.0.0 - November, 9 2021
 
+* Add - Demo screenshot.
 * Tweak - Update Intelephense to version 1.8.0.
 
 ## Version 1.0.0 - November, 9 2021